requirement.txt
appdirs==1.4.4
audioread==3.0.0
av==10.0.0
blobfile==2.0.1
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==39.0.1
decorator==4.4.2
docker-pycreds==0.4.0
einops==0.6.0
einops-exts==0.0.4
filelock==3.9.0
flit_core==3.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
lazy_loader==0.1
librosa==0.10.0
llvmlite==0.39.1
lpips==0.1.4
lxml==4.9.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
msgpack==1.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.0
pooch==1.7.0
proglog==0.1.10
protobuf==4.22.1
psutil==5.9.4
pycparser==2.21
pycryptodomex==3.17
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.11
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.1
scipy==1.10.1
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.4
termcolor==2.2.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.14
visdom==0.2.4
wandb==0.13.11
wcwidth==0.2.6
websocket-client==1.5.1
wheel==0.38.4
xmltodict==0.13.0
zipp==3.15.0